Build the sync unit tests:
 - Remove old main(argc, argv) from syncable_unittest.
 - Since the old notifier target had only one unittest,
merge it with the syncer tests.
 - Remove listener_unittest.cc entirely.
 - Don't compile natserver_main.cc, which has another main()
 - Use a test main from the test_support_unit library.  This
   has the downside of pulling in a lot of deps we don't need,
   but it gives us a run_all_unittests.cc.
 - Fix a few compile errors in test code.

TEST=unit tests pass
BUG=none

Review URL: http://codereview.chromium.org/205021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26358 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/sync/util/event_sys_unittest.cc
@@ -153,7 +153,7 @@ class ThreadTester : public EventListener<TestEvent> {
   }
   ~ThreadTester() {
     pair_->event_channel()->RemoveListener(this);
-    for (int i = 0; i < threads_.size(); i++) {
+    for (size_t i = 0; i < threads_.size(); i++) {
       CHECK(pthread_join(threads_[i].thread, NULL) == 0);
       delete threads_[i].completed;
     }
@@ -230,7 +230,7 @@ class ThreadTester : public EventListener<TestEvent> {
     sleep(1);
 #endif
 
-    for (int i = 0; i < threads_.size(); i++) {
+    for (size_t i = 0; i < threads_.size(); i++) {
       if (*(threads_[i].completed))
         LOG(FATAL) << "A test thread exited too early.";
     }
